AccountManager / testing
+ AccountManager::createCharacter and deleteCharacter are working again.
+ Time now the correct type (i64).
+ Gender is sanity checked during Character creation now. Deity is not
yet implemented.
+ New tests for create / delete character.

+++ world/Utility.cpp
@@ -121,8 +121,8 @@ String Utility::getRandomName() {
 	return String(rndname);
 }
 
-int32 Utility::Time::now() {
-	return static_cast<int32>(std::time(nullptr));
+i64 Utility::Time::now() {
+	return std::time(nullptr);
 }
 
 const String& Utility::Skills::fromID(const uint32 pSkillID) {
